# spring-boot-openapi-generator/src/main/java/org/openapitools/api/CreditLimitApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.4.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package org.openapitools.api;

import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import javax.validation.Valid;
import javax.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import javax.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2024-03-27T14:56:01.389439+05:30[Asia/Kolkata]", comments = "Generator version: 7.4.0")
@Validated
@Tag(name = "credit-limit", description = "the credit-limit API")
public interface CreditLimitApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    /**
     * PUT /credit-limit/decrease-limit : Decreasing credit limit
     *
     * @param body  (optional)
     * @return Credit limit reduced successfully, customer informed (status code 200)
     */
    @Operation(
        operationId = "creditLimitDecreaseLimitPut",
        summary = "Decreasing credit limit",
        responses = {
            @ApiResponse(responseCode = "200", description = "Credit limit reduced successfully, customer informed")
        }
    )
    @RequestMapping(
        method = RequestMethod.PUT,
        value = "/credit-limit/decrease-limit",
        consumes = { "application/json" }
    )
    
    default ResponseEntity<Void> creditLimitDecreaseLimitPut(
        @Parameter(name = "body", description = "") @Valid @RequestBody(required = false) Object body
    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * PUT /credit-limit/increase-limit : Increasing credit limit
     *
     * @param body  (optional)
     * @return Credit limit increased successfully, customer informed (status code 200)
     */
    @Operation(
        operationId = "creditLimitIncreaseLimitPut",
        summary = "Increasing credit limit",
        responses = {
            @ApiResponse(responseCode = "200", description = "Credit limit increased successfully, customer informed")
        }
    )
    @RequestMapping(
        method = RequestMethod.PUT,
        value = "/credit-limit/increase-limit",
        consumes = { "application/json" }
    )
    
    default ResponseEntity<Void> creditLimitIncreaseLimitPut(
        @Parameter(name = "body", description = "") @Valid @RequestBody(required = false) Object body
    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}