# spring-boot-openapi-generator/src/main/java/org/openapitools/api/RewardPointsApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.4.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package org.openapitools.api;

import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import javax.validation.Valid;
import javax.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import javax.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2024-03-27T14:56:01.389439+05:30[Asia/Kolkata]", comments = "Generator version: 7.4.0")
@Validated
@Tag(name = "reward-points", description = "the reward-points API")
public interface RewardPointsApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    /**
     * POST /reward-points/add-points : Adding reward points
     *
     * @param body  (optional)
     * @return Rewards points added successfully (status code 201)
     */
    @Operation(
        operationId = "rewardPointsAddPointsPost",
        summary = "Adding reward points",
        responses = {
            @ApiResponse(responseCode = "201", description = "Rewards points added successfully")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/reward-points/add-points",
        consumes = { "application/json" }
    )
    
    default ResponseEntity<Void> rewardPointsAddPointsPost(
        @Parameter(name = "body", description = "") @Valid @RequestBody(required = false) Object body
    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * PUT /reward-points/redeem-points : Redeeming reward points
     *
     * @param body  (optional)
     * @return Rewards redeemed successfully, points updated (status code 200)
     */
    @Operation(
        operationId = "rewardPointsRedeemPointsPut",
        summary = "Redeeming reward points",
        responses = {
            @ApiResponse(responseCode = "200", description = "Rewards redeemed successfully, points updated")
        }
    )
    @RequestMapping(
        method = RequestMethod.PUT,
        value = "/reward-points/redeem-points",
        consumes = { "application/json" }
    )
    
    default ResponseEntity<Void> rewardPointsRedeemPointsPut(
        @Parameter(name = "body", description = "") @Valid @RequestBody(required = false) Object body
    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}